# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comment/discussion/{discussionId}": {
            "get": {
                "tags": [
                    "Discussion's Comment"
                ],
                "summary": "Get specific discussion's comments by ID",
                "operationId": "463b2371bee326b9cd078852cf422e6d",
                "parameters": [
                    {
                        "name": "discussionId",
                        "in": "path",
                        "description": "Id of the discussion",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get comments of this discussion"
                    },
                    "500": {
                        "description": "Cannot get comments of this discussion!"
                    }
                }
            }
        },
        "/api/community": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Search for communities by filters",
                "operationId": "42238015ea4db8f0efbf8bf750bf4390",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the community to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "visibility",
                        "in": "query",
                        "description": "Visibility of the community: all, public, or private",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "all",
                                "public",
                                "private"
                            ]
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "User's role in the community: admin or member",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "member"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Paginate",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search community success"
                    },
                    "500": {
                        "description": "Cannot search community!"
                    }
                }
            }
        },
        "/api/community/route/{route}": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Get specific community by route",
                "operationId": "b302fafb2c460bb6348442ec6c5dede7",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route of the community",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1mmee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get community success"
                    },
                    "500": {
                        "description": "Cannot get community!"
                    }
                }
            }
        },
        "/api/community/{route}/permission/home": {
            "get": {
                "tags": [
                    "Community"
                ],
                "summary": "Check community permission",
                "operationId": "8d1df6a8dbdb8f9ec4bc683b2607d680",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Route of the community",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "1mmee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Community permission successfully"
                    },
                    "500": {
                        "description": "Cannot get Community permission!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}